requirements.txt
flask==2.3.3
opencv-python==4.8.0.76
deepface==0.0.79
numpy==1.24.3
paho-mqtt==1.6.1
python-dotenv==1.0.0
schedule==1.2.0